moved the results structs to the results package. Allows more flexibility and no import cycles
