+++ analyze/analyze.go
@@ -4,6 +4,7 @@ package analyze
 import (
 	"context"
 	"fmt"
+	"github.com/boostsecurityio/poutine/results"
 	"os"
 	"sync"
 	"time"
@@ -286,7 +287,7 @@ func (a *Analyzer) AnalyzeLocalRepo(ctx context.Context, repoPath string) error 
 }
 
 type Formatter interface {
-	Format(ctx context.Context, report *opa.FindingsResult, packages []*models.PackageInsights) error
+	Format(ctx context.Context, report *results.FindingsResult, packages []*models.PackageInsights) error
 }
 
 func (a *Analyzer) finalizeAnalysis(ctx context.Context, inventory *scanner.Inventory) error {
