+++ scanner/inventory_test.go
@@ -2,6 +2,7 @@ package scanner
 
 import (
 	"context"
+	"github.com/boostsecurityio/poutine/results"
 	"testing"
 
 	"github.com/boostsecurityio/poutine/models"
@@ -68,11 +69,11 @@ func TestFindings(t *testing.T) {
 	err := i.AddScanPackage(context.Background(), *pkg, "testdata")
 	assert.Nil(t, err)
 
-	results, err := i.Findings(context.Background())
+	analysisResults, err := i.Findings(context.Background())
 	assert.Nil(t, err)
 
 	rule_ids := []string{}
-	for _, r := range results.Rules {
+	for _, r := range analysisResults.Rules {
 		rule_ids = append(rule_ids, r.Id)
 	}
 
@@ -91,11 +92,11 @@ func TestFindings(t *testing.T) {
 		"unverified_script_exec",
 	})
 
-	findings := []opa.Finding{
+	findings := []results.Finding{
 		{
 			RuleId: "debug_enabled",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/debug_enabled_valid.yml",
 				Details: "ACTIONS_RUNNER_DEBUG",
 			},
@@ -103,7 +104,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "debug_enabled",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Job:     "build",
 				Path:    ".github/workflows/debug_enabled_valid.yml",
 				Line:    9,
@@ -113,7 +114,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "debug_enabled",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Job:     "build",
 				Path:    ".github/workflows/debug_enabled_valid.yml",
 				Step:    "0",
@@ -124,7 +125,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "injection",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Job:     "build",
 				Path:    ".github/workflows/valid.yml",
 				Step:    "1",
@@ -135,7 +136,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "injection",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Job:     "build",
 				Path:    ".github/workflows/valid.yml",
 				Step:    "7",
@@ -146,7 +147,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "known_vulnerability_in_build_component",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    "composite/action.yml",
 				OsvId:   "GHSA-4mgv-m5cm-f9h7",
 				Step:    "2",
@@ -157,7 +158,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "known_vulnerability_in_build_component",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/valid.yml",
 				Job:     "build",
 				Step:    "5",
@@ -169,7 +170,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "known_vulnerability_in_build_component",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/valid.yml",
 				Job:     "build",
 				Step:    "6",
@@ -181,7 +182,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "untrusted_checkout_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/valid.yml",
 				Line:    30,
 				Details: "Detected usage of `npm`",
@@ -190,7 +191,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "untrusted_checkout_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/valid.yml",
 				Line:    56,
 				Details: "Detected usage the GitHub Action `bridgecrewio/checkov-action`",
@@ -199,7 +200,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "untrusted_checkout_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/valid.yml",
 				Line:    60,
 				Details: "Detected usage of `pre-commit`",
@@ -208,7 +209,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "untrusted_checkout_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/workflow_run_valid.yml",
 				Line:    13,
 				Details: "Detected usage of `npm`",
@@ -217,28 +218,28 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "default_permissions_on_risky_events",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path: ".github/workflows/valid.yml",
 			},
 		},
 		{
 			RuleId: "unpinnable_action",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path: "action.yml",
 			},
 		},
 		{
 			RuleId: "unpinnable_action",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path: "composite/action.yml",
 			},
 		},
 		{
 			RuleId: "pr_runs_on_self_hosted",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/valid.yml",
 				Job:     "build",
 				Line:    9,
@@ -247,7 +248,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "pr_runs_on_self_hosted",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/allowed_pr_runner.yml",
 				Job:     "group",
 				Line:    13,
@@ -256,7 +257,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "pr_runs_on_self_hosted",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/allowed_pr_runner.yml",
 				Job:     "labels",
 				Line:    19,
@@ -265,14 +266,14 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "github_action_from_unverified_creator_used",
 			Purl:   "pkg:githubactions/kartverket/github-workflows",
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Details: "Used in 1 repo(s)",
 			},
 		},
 		{
 			RuleId: "injection",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Job:     "build",
 				Path:    ".github/workflows/valid.yml",
 				Step:    "8",
@@ -283,7 +284,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "injection",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".gitlab-ci.yml",
 				Job:     "default.before_script[0]",
 				Details: "Sources: inputs.gem_name",
@@ -293,7 +294,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "debug_enabled",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".gitlab-ci.yml",
 				Details: "CI_DEBUG_SERVICES CI_DEBUG_TRACE",
 			},
@@ -301,7 +302,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "job_all_secrets",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path: ".github/workflows/secrets.yaml",
 				Line: 4,
 				Job:  "matrix",
@@ -310,7 +311,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "job_all_secrets",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path: ".github/workflows/secrets.yaml",
 				Line: 16,
 				Job:  "json",
@@ -319,7 +320,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "if_always_true",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path: "composite/action.yml",
 				Line: 17,
 				Step: "3",
@@ -328,7 +329,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "unverified_script_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/valid.yml",
 				Line:    70,
 				Job:     "build",
@@ -339,7 +340,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "unverified_script_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".github/workflows/valid.yml",
 				Line:    75,
 				Job:     "build",
@@ -350,7 +351,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "unverified_script_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    "azure-pipelines-1.yml",
 				Line:    8,
 				Step:    "2",
@@ -360,7 +361,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "injection",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".azure-pipelines.yml",
 				Line:    14,
 				Job:     "build",
@@ -371,7 +372,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "debug_enabled",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".azure-pipelines.yml",
 				Line:    0,
 				Job:     "",
@@ -382,7 +383,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "untrusted_checkout_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    "azure-pipelines-2.yml",
 				Line:    14,
 				Job:     "",
@@ -393,7 +394,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "untrusted_checkout_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    "azure-pipelines-4.yml",
 				Line:    11,
 				Job:     "",
@@ -404,7 +405,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "untrusted_checkout_exec",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".tekton/pipeline-as-code-tekton.yml",
 				Line:    43,
 				Job:     "vale",
@@ -415,7 +416,7 @@ func TestFindings(t *testing.T) {
 		{
 			RuleId: "injection",
 			Purl:   purl,
-			Meta: opa.FindingMeta{
+			Meta: results.FindingMeta{
 				Path:    ".tekton/pipeline-as-code-tekton.yml",
 				Line:    45,
 				Job:     "vale",
@@ -425,8 +426,8 @@ func TestFindings(t *testing.T) {
 		},
 	}
 
-	assert.Equal(t, len(findings), len(results.Findings))
-	assert.ElementsMatch(t, findings, results.Findings)
+	assert.Equal(t, len(findings), len(analysisResults.Findings))
+	assert.ElementsMatch(t, findings, analysisResults.Findings)
 }
 
 func TestSkipRule(t *testing.T) {
@@ -445,11 +446,11 @@ func TestSkipRule(t *testing.T) {
 	err := i.AddScanPackage(ctx, *pkg, "testdata")
 	assert.Nil(t, err)
 
-	results, err := i.Findings(context.Background())
+	analysisResults, err := i.Findings(context.Background())
 	assert.Nil(t, err)
 
 	rule_ids := []string{}
-	for _, r := range results.Findings {
+	for _, r := range analysisResults.Findings {
 		rule_ids = append(rule_ids, r.RuleId)
 	}
 
@@ -464,11 +465,11 @@ func TestSkipRule(t *testing.T) {
 	})
 	assert.NoError(t, err)
 
-	results, err = i.Findings(context.Background())
+	analysisResults, err = i.Findings(context.Background())
 	assert.Nil(t, err)
 
 	rule_ids = []string{}
-	for _, r := range results.Findings {
+	for _, r := range analysisResults.Findings {
 		rule_ids = append(rule_ids, r.RuleId)
 	}
